CCPipelines.jl

A simple package to aid in setting up data pipelines for my work at Cayuga Centers.

CCPipelines.a — Method

a(b)

Returns b plus 1.

CCPipelines.a — Method

a(b::Int)

When b is an Int, returns b plus 2.